pyproject.toml
[tool.poetry]
name = "packyak"
version = "0.1.0"
description = "Infrastructure for AI applications and machine learning pipelines"
authors = ["sam <sam@eventual.ai>"]
readme = "README.md"


[tool.poetry.dependencies]
types-aiobotocore = { extras = ["essential"], version = "^2.9.0" }
types-boto3 = "^1.0.2"

aioboto3 = "^12.1.0"
aiobotocore = { version = "2.8.0", extras = ["boto3"] }
aws-lambda-typing = "^2.18.0"
boto3-stubs = { extras = ["essential"], version = "^1.34.15" }
pandas = "^2.1.4"
pandera = "^0.18.0"
pydantic = "^2.5.3"
python = "3.12.1"
pandas-stubs = "^2.1.4.231227"

[tool.poetry.group.dev.dependencies]
matplotlib = "^3.8.2"
pylint = "^3.0.3"
pyright = "^1.1.345"
requests = "^2.31.0"
ruff = "^0.1.11"
streamlit = "^1.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
ignore = ["E402", "F401"]
